+++ freekick/api/match.py
@@ -1,23 +1,34 @@
 """API that calls the match_predictor service"""
 
-from flask import Blueprint, jsonify, request
+from flask_restful import Resource, fields, marshal_with, reqparse
 
 from freekick.service import predict_match
 
-match_route = Blueprint("match_route", __name__, url_prefix="/api")
+resource_fields = {
+    "home_team": fields.String,
+    "away_team": fields.String,
+    "predicted_winner": fields.String,
+}
+
+post_parser = reqparse.RequestParser()
+post_parser.add_argument("home", type=str, required=True, help="Home team code")
+post_parser.add_argument("away", type=str, required=True, help="Away team code")
+post_parser.add_argument("league", type=str, required=True, help="league code")
+post_parser.add_argument(
+    "attendance", type=float, help="Approximate. Average attendance"
+)
 
 
-@match_route.route("/match", methods=["GET", "POST"])
-def match():
-    request_data = request.get_json()
-    home_team = request_data["home"].replace(" ", "-")
-    away_team = request_data["away"].replace(" ", "-")
-    league = request_data["league"]
-    attendance = request_data["attendance"]
-    prediction = predict_match(
-        league=league,  # should we create a League enum here or have predict_match create the enum?
-        home_team=home_team,
-        away_team=away_team,
-        attendance=attendance,
-    )
-    return jsonify(prediction)
+class Match(Resource):
+    @marshal_with(resource_fields)
+    def post(self):
+        args = post_parser.parse_args(strict=True)
+        match_dto = predict_match(
+            league=args[
+                "league"
+            ],  # should we create a League enum here or have predict_match create the enum?
+            home_team=args["home"].replace(" ", "-"),
+            away_team=args["away"].replace(" ", "-"),
+            attendance=args["attendance"],
+        )
+        return match_dto