+++ freekick/service/__init__.py
@@ -1,4 +1,5 @@
 from .match_day_predictor import predict_match_day
 from .match_predictor import predict_match
+from .util import MatchDTO
 
-__all__ = ["predict_match_day", "predict_match"]
+__all__ = ["predict_match_day", "predict_match", "MatchDTO"]