+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Version 0.2
 - Whats new in version 0.2.0
 
   - _LEAGUES has been renamed to League and is now an Enum.
+  - Migrate handlers from Flask Blueprint to flask-restful
 
 Version 0.1
 -----------